+++ src/main/java/de/samply/frontend/FrontendConfiguration.java
@@ -1,6 +1,7 @@
 package de.samply.frontend;
 
 import de.samply.app.ProjectManagerConst;
+import jakarta.annotation.PostConstruct;
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
@@ -16,6 +17,17 @@ public class FrontendConfiguration {
 
     private String baseUrl;
     private Map<String, String> sites = new HashMap<>();
+
+    @PostConstruct
+    private void postConstruct() {
+        Map<String, String> tempSites = new HashMap<>();
+        sites.forEach((siteId, sitePath) -> tempSites.put(replaceHyphen(siteId), sitePath));
+        sites = tempSites;
+    }
+
+    private String replaceHyphen(String var) {
+        return var.replace(ProjectManagerConst.HYPHEN, "-");
+    }
 
     public Optional<String> getSitePath(String site) {
         return Optional.ofNullable(sites.get(site));